Create a function for fetching the parameter movie
